+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - New plugin `sri`.
 - Improved plugin docs with links to the online documentation.
 - New functions `nav.nextPage()` and `nav.previousPage()` for `nav` plugin.
+- New sort options `asc-locale` and `desc-locale`.
 
 ### Removed
 - `cms.ts` file.